Flow changes, qr code broken
No more nav drawer in the main screen (useless for it to be there
anyways). Timeout field is removed from the add server screen and will
only be shown when editing a server. Timeout default was also raised
from 3 seconds to 15, maybe do a wifi timeout and a 3g/4g timeout at a
later time?

QR code parsing is broken, I have no idea how to fix it.