Add temp FERemoteLib, Start Service
Added new module to allow connection to FERemote, this is to be replaced
with a jar file once we have official builds.
A new Service that doesn't do much yet was added.

Basically a non-commit because I need to move my computer and make sure
I don't lose any data.