Implement chat log
Updated FERemoteLib (this needs to be moved to maven at one point)

A working push chat (receiving only) can be seen in the chat section of
the app. It will keep receiving messages as long as the service running
in the background is connected to the app, this may or may not cause a
lot of battery drainage. Needs to be tested.